--ammend

+++ tests/integration/suite/daprd/metrics/actors/actors.go
@@ -101,6 +101,6 @@ func (a *actors) Run(t *testing.T, ctx context.Context) {
 
 	assert.EventuallyWithT(t, func(c *assert.CollectT) {
 		metrics := a.daprd.Metrics(c, ctx).All()
-		assert.Equal(c, 5, int(a.triggered.Load()), int64(metrics["dapr_runtime_actor_reminders_fired_total"]))
+		assert.Equal(c, 5, int(a.triggered.Load()), int(metrics["dapr_runtime_actor_reminders_fired_total"]))
 	}, time.Second*20, 10*time.Millisecond)
 }
